requirements.txt
accelerate==0.34.0
auraloss==0.4.0
braceexpand==0.1.7
clean_fid==0.1.35
clip==1.0
deepdiff==8.4.1
diffusers==0.30.2
ftfy==6.3.1
h5py==3.12.1
horovod==0.28.1
librosa==0.10.2.post1
matplotlib==3.10.1
numpy==2.2.4
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
packaging==24.2
pandas==2.2.3
Pillow==11.1.0
PySoundFile==0.9.0.post1
regex==2024.7.24
scikit_learn==1.6.1
scipy==1.15.2
simplejpeg==1.7.6
soundfile==0.12.1
timm==0.6.7
torch==2.3.0
torchaudio==2.3.0
torchlibrosa==0.1.0
torchsort==0.1.9
torchvision==0.18.0+cu121
tqdm==4.66.5
transformers==4.44.2
wandb==0.19.8
wav2clip==0.1.0
webdataset==0.2.100
wget==3.2
